break out homepage component and css into their own files
