+++ index.html
@@ -8,93 +8,7 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/pdf.js/2.6.347/pdf.min.js" integrity="sha512-Z8CqofpIcnJN80feS2uccz+pXWgZzeKxDsDNMD/dJ6997/LSRY+W4NmEt9acwR+Gt9OHN0kkI1CTianCwoqcjQ==" crossorigin="anonymous"></script>
     <link rel="preconnect" href="https://fonts.gstatic.com">
     <link href="https://fonts.googleapis.com/css2?family=Open+Sans&display=swap" rel="stylesheet">
-    <style>
-        body {
-            text-align: center;
-            font-family: 'Open Sans', sans-serif;
-        }
-        
-        canvas {
-            border: 1px solid #000;
-        }
-        
-        button {
-            margin-left: 1%;
-            margin-right: 1%;
-        }
-        
-        ul {
-            list-style-type: none;
-            padding: 0;
-        }
-        
-        li {
-            margin-top: 10px;
-            margin-bottom: 10px;
-        }
-        
-        .about {
-            font-size: 16px;
-            text-align: left;
-        }
-        
-        .aboutHeader:hover {
-            cursor: pointer;
-        }
-        
-        #buttons {
-            margin: 3%;
-        }
-        
-        #container {
-            display: grid;
-            grid-template-columns: repeat(auto-fit, minmax(500px, 1fr));
-        }
-        
-        #openScoreInTab {
-            display: block;
-            margin: 1% auto;
-        }
-        
-        #theCanvas {
-            margin-top: 1%;
-        }
-        
-        table {
-            margin: 0 auto;
-            width: 80%;
-            border-collapse: collapse
-        }
-        
-        span {
-            font-size: 12px;
-        }
-        
-        a {
-            text-decoration: none;
-        }
-
-        a:hover {
-            text-decoration: underline;
-        }
-        
-        .checkbox {
-            margin-right: 3%;
-        }
-        
-        #instrumentControls {
-            margin-top: 1%;
-        }
-        
-        #notesContainer {
-            text-align: left;
-            padding: 2%;
-        }
-        
-        #notesHeader, #instrumentName {
-            font-weight: bold;
-        }
-    </style>
+    <link href="style.css" rel="stylesheet">
 </head>
 
 <body>
